Gem upgradeathon
In the process of upgrading Rails to 3.2.13, all of the gem dependencies
have now been made more specific to help prevent the possibility of a
subsequent upgrade breaking things.

Of course, the only real way to fight this is to write really good
automated tests. Those will hopefully be coming soon.