Move devise views from devise -> users view directory
We originally went with the default `devise` view directory because it
was quicker to get up and running with. Now that we have time to come
back to this, this change moves all devise-related views (which all
happen to be related to the user at this point in time) from the
app/views/devise directory to the app/views/users directory.

This will make it easier to identify where to go to edit these views in
the future.