Allow customization of ticket pricing and limits
Adds the ability for site admins to set the prices of adult/kid tickets
and cabins.

The large sweeping changes on various pages that resulted from this new
functionality are simply incorporating those values into the various
places they are used.

This also adds a heap of specs to the Event model to ensure the
validations are working as expected.