Add `volunteer_shifts` field to ticket request
This field asks users how many shifts they are willing to volunteer for
when signing up for tickets. The current thinking is that this will not
be rigorously enforced, but will be used to determine priority of who
gets which tickets, and for estimating how many shifts an event will be
able to fill.