Manually stop and start unicorn on server restart
The zero-downtime restart of unicorn is proving problematic, where the
master process does not seem to respect the USR2 signal in certain
situations.

Since we don't have the level of traffic that justifies zero-downtime
deploys, for now we just manually stop and start the server when issuing
the "restart" command.