Polish event volunteer/payment flow
In the process of manually QA-ing this flow, a found a number of bits of
weirdness popped up. This attempts to address a large number of them.

Main takeaway is that users are now taken to the staff sign-up page
before they can pay for their ticket. This provides a way to "nudge"
people into volunteering for shifts.