Add custom price rule support for kids tickets
The PriceRule model was added to enable support for defining "custom"
prices for a certain number of tickets. These still need to be created
from the command line, as making a UI to create them is probably not the
best use of time at this point.

Also added support for displaying the cost of adult, kid, and cabin
portion of the ticket price as user changes the values.

Closes #2