Add quick job-creation wizard
Previously, in order to create a job with many shifts the event
admin would have to create the job and then individually create the time
slots for each shift. This is time-consuming, so an automated method
where you could specify the job start/end times and the length of the
shift was in order.

This adds that functionality to the job form (but only when creating a
job--this cannot be edited).

In the process, it was realized that creating individual `TimeSlot`
records isn't the smartest way of doing things if we _know_ that shifts
will alway be of equal length. In the future, we should remove the
concept of a `TimeSlot` entirely, and instead just store the shift
length information on the `Job` model, and add the time the user
volunteered for to the `Shift` model. From there we can dynamically
calculate which time slots have been filled, without requiring as many
database records.

Closes #1.