Load Stripe configuration from YAML file
Previously, Stripe integration was hard-coded to use the "test"
secret/public API keys. Now that we're preparing for actually taking
credit cards, load the configuration from a `config/stripe.yml`.

We don't check this file in to source control because it contains
sensitive information. Thus an example file is provided under
`config/stripe.yml.example` which works out of the box, but only uses
the development Stripe API keys. The production keys will only exist on
the production server.