Make batch time slot creation easier
Adds support for creating the next time slot for a job based on the
previous time slot by default. This will make entry after the job has
already been created (as opposed to at the time the wizard was run) a
little easier, which is important when we're working with live data.