Add "completed" status to ticket requests
After a user has paid for their ticket, we need an additional status to
indicate they their request is "complete" without requiring that a
payment have been made (since some tickets can be $0).

This adds a "completed" status to ticket requests, and changes the
behaviour of ticket approval to mark the ticket as complete if it is a
$0 ticket.

In the process, the ticket request index was refactored and optimized to
use `includes` to significantly reduce the number of queries by removing
N+1 queries.