Add Mandrill inbound email processing
This alleviates the need for us to manage our own MTA setup. It also
opens up the possibility of one day allowing ticket requests to reply
directly to emails to get in touch with event coordinators.

For now, this exposes an API endpoint that Mandrill can hit when it
receives an email on our behalf. This endpoint extracts the relevant
information from the email and then forwards it to my email (eventually
we'll have this forward to the event coordinator's email or something
similar).

In order to support the web hook, I had to remove `force_ssl` from the
superclass ApplicationController so I could individually disable it in
the WebHookController (apparently Mandrill doesn't like our SSL
certificate). Since I don't have time to fix the certificate today, I
simply duplicated `force_ssl` for all other controllers except the
WebHookController.