Wire up Mandrill as SMTP relay
In an effort to reduce the likelihood that our email gets marked as
spam, we're going to delegate our email sending to Mandrill. This also
has the benefit of providing us with tracking and open rates, and for
our volume of email should be free.

Includes an example YAML file for documentation purposes.