Add schedule calendar to shift signup page
This is more user-friendly than a basic list format, as it allows you to
easily compare who else is working on other shifts during a particular
time period.

In order to support this, we had to enforce an assumption that all
shifts start/end on the 30 minute mark. A shift can still end at the
quarter mark, but it will be rounded down when displayed in the calendar
view, otherwise we would have problems rendering using table cells
coupled with `rowspan`s.