Change product name to Helping Culture
As we move towards a product that is intended to take on other events,
we need to displace itself from the CloudWatch name. This represents the
first step in that direction (although the name may very well change in
future).

We haven't done a full code rename yet since we're not sure of this
direction. We'll see what happens in the coming months.