Add error reporting support via Sentry
In order to be alerted when problems occur, we add error reporting using
the sentry-raven gem with a hosted Sentry instance. Right now it's free
with very low limits, but if we need to upgrade it's not too expensive
to do so, and it's a good service. Right now it'll do a good job of
sending emails when something breaks.

This installs a specific git revision for the gem since it has a fix for
an SSL verification bug in version 0.4.8 (0.4.9 hadn't been released at
this time of writing).

The reporting is only enabled on production environments, and the
sentry.yml file is only loaded in that environment.