Fix photo upload when uploading no file
If a user uploaded a photo, and then attempted to modify another
attribute at a later time without modifying the photo, the old photo
would end up accidentally overwritten with a new (non-existent) file.

This was caused by the fact that we were generating a new secure token
for the file because we weren't attempting to detect whether or not the
file had actually changed. Adding this check solves the problem.

In the process, we also changed the location of the cache directory so
that it will work out-of-the-box with the development environment. We'll
have to follow this up with a fix that periodically cleans out the cache
directory so it doesn't fill up.