Use nested form fields for user
Previously, the ticket request form was hacked to create a user from
some ad hoc form fields on the page. This isn't ideal, as Rails actually
provides a convenient nested form field solution to deal with this
exact situation. This commit changes the implementation to use that
feature (specifically, the `accept_nested_attributes_for` DSL).

In the process, the requirement that a user can only create one ticket
request has been removed, and we've also fixed a vulnerability where a
user could specify which user the ticket request belonged to without
being that user.