Allow events to accept donations
Events can now specify whether they would like to accept donations and
if so receive donations on the payment page when users pay for their
ticket. The donation form displays 5 options ($0, $10, $25, $50, and
$100).

This also adds information to the payment screen to show the total
amount paid and credit card information. There was also some
miscellaneous changes to wording on the event edit form.