Fix broken ticket request spec
Commit 319cf6d7 removed a validation that ensured the existence of a
user for a ticket request. That commit removed the uniqueness constraint
of users being limited to one ticket request per event.

The commit didn't mention the removal of this existence validation, so I
think it was done in error. I've added it back in and fixed a spec that
was was broken in the process.