Don't set date on datetimepicker when date null
When a null date was passed to the `datetimepicker` helper, the JS would
render a picker with date at the beginning of the UNIX epoch, rather
than no date.

This was due to us initializing a date with a null value, which is
easily fixed by checking that `date` is not null.