Update 500 page to be a little more friendly
What better thing would there be to see when a page errors than a funny
cat video?

This has been implemented as a standalone file so that even in the most
dire of circumstances (e.g. asset precompilation broke for some reason)
this page would still work since it's served by NGINX.

Idea stolen from a popular website who shall not be named.