Fix existence validators
The upgrade to Rails 4 broke all existence validators due to the removal
of the `primary_key_name` method from `ActiveRecord`. A commit was
recently merged in the `validates_existence` gem repo, but has yet to be
deployed to Rubygems.

For now, specify the exact repo and revision so we can use the fix now.