Add auth tokens to request confirmation email
This will save users from having to enter their email/password when
purchasing tickets (which would be annoying).

There are a number of limitations with this (e.g. generating a new auth
token overwrites any previously existing ones) but for the flow that
we're currently supporting it's adequate. Would love to implement a
ctoken approach if time permitted.