Allow approval emails to be re-sent
Add a button to the ticket request index page allowing approval emails
to be resent to ticket requests "awaiting payment".

In its current form this is a bit dangerous as multiple people can
clickk the button and effectively spam the user. Once time permits
we'll probably change this to a "Remind" button and add a `reminded_at`
column so we can remind the user at most once a day (or something along
those lines).