Quiet resend approval button
This isn't a common action, so it should be a smaller link instead of a
button.

While here we also added a clause to only show the link if the ticket
request isn't completed (for requests which have no payment but are
complete, e.g. free requests, we don't want to show the button).