Switch email sending to use GMail
Mandrill removed their free tier and raised prices, so we're not using
them anymore.

GMail allows us to send up to 500 emails per day. This is well within
the volume we send, so use the fnftickets@gmail.com account for this
purpose.

While here I removed some dead code and moved the SMTP configuration
into config/application.rb so that it will load it if it finds it,
regardless of environment (that way we can test in dev).