Add support for purchasing Early Arrival/Late Departure pass
This quickly blew up into more work than expected.

We want to be able to sell EA/LD passes separately from tickets (since
they are not refundable. This required setting up a separate Stripe
charge, which required the creation of a new model and associated views.
Downloading a CSV is also possible so we can send it to the team
managing EA/LD passes.

This includes a nasty hack so our payments JS works for both the tickets
and EALD payment forms. I don't have any good excuse for this blight
except my lack of enthusiasm for putting more time into this project.