Remove validates_existence gem
This gem stopped being supported after Rails 3 and was causing the
display of endless lines like the following when running tests:

    DEPRECATION WARNING: You are passing an instance of
    ActiveRecord::Base to `exists?`

Fix it by removing the validations. These don't provide a ton of value
(but were helpful when we were first building the system). Now our test
output is much more legible.