Add support for specifying guest list
We want to require ticket purchasers to specify their guest list
_before_ they purchase a ticket. This will ensure that we can hold them
accountable to the guests they intend to bring.

The change to the ticket request form is minor even though it looks like
much more--we had to set it up to display only the modifiable guest list
if the user was editing a ticket request, otherwise render the original
form.

The approach here is quick and dirty to minimize effort. If it becomes a
burden to maintain we'll suffer the consequences later.