Prevent two payments to the same ticket request
We had a situation where someone was double charged. This happened
because they refreshed or otherwise re-clicked the button some time
after they had already clicked it (or perhaps the network connection got
severed and their client didn't detect).

This isn't a perfect solution, but given in both situations this
occurred the time difference between each of the Payment records being
created was on the order of many seconds, it seems like it's a simple
way to avoid this issue.